USB HAL: default: Fix incorrect function pointer cast
A pointer to uevent_event() is stored in epoll_event::data::ptr
which is later invoked after the poll returns. The function
pointer is currently cast with a signature that doesn't quite
match the actual function, namely the first parameter should
be uint32_t, not int. Fix the potential control flow integrity
vulnerability.

CRs-Fixed: 2324304
Change-Id: I42b926efd09b6978746b9b14138bafbb3921677d